app/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==6.2.1
PyYAML==5.3.1
six==1.15.0
torch==1.3.1
torchvision==0.4.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
